Limit TypeScript to minor version 5.1.x
This is to keep the version of ESLint we use happy, which in turn is because
we can’t advance to a later ESLint because the AirBnb ESLint rulesets have
not been updated for that (and seem unlikely to be).

It would be nice to use e.g. “Standard”, but that would be a lot of work at
present.