Export valToJs and use it in tests
Test that the test result is equal to the expected result using POJO
comparison (`deepEqual`), not text.

Additionally, for each test, ensure that compiling the input and
reconverting to JS gives the input again (again comparing the two as POJOs).

Fix a couple of bugs in serialization as a result, for `loop` and `if`.